Add files via upload
-Debugged some issues and enhanced the code
-Added user Authentication for Secure CRUD operations

+++ main.sh
@@ -3,10 +3,15 @@
 installFlag=0 
 if ! command -v kdialog &> /dev/null
 then
-   echo "a key command was not found on your system ,Please Install kdialog and rerun the script"
+   echo "a key command was not found on your system, Please Install kdialog and rerun the script"
    exit
 fi
 
+# Check if user is logged in
+if [ -z "$loggedInUser" ]; then
+    echo "You must log in first."
+    exit 1
+fi
 
 clear
 kdialog --menu "Operations Menu" "1" "Create Table" "2" "List Tables" "3" "Drop Table" "4" "Insert into Table" "5" "Select From Table" "6" "Delete From Table" "7" "Update Table" "0" "Exit" > out
@@ -33,3 +38,4 @@ done
 if [ -f out ];then
 rm out
 fi
+
